Transport review (#35)
* Refs #3441. Added API strucure and examples.

* Refs #3446. First approach.

* Refs #3446. Make basic discovery implementation.

* Refs #3446. Added header parse

* Refs #3446. Adapted to rebase develop changes.

* Generic ip initialization udp. Reduce stack at udp connection.

* Completed funtionality of discovery. It is necessary a great build system mode for the transport OS dependent.

* Refs #3446 Added build system.

* Refs #3446. Fixed bug at serialization GET INFO message. Updated log for discovery.

* Refs #3446. Modified array to IP address convertion.

* Slight changes

* Post rebase changes.

* Refs #3447. Updated Cross serlization tests to support discovery messages.

* Refs #3441. Change API.

* Refs #3447. Added integration test. First approach.

* Addapt integration test with new xml representation.

* Refs #3551. Added support for TCP in interaction test.

* Default create/delete session attempts reduced.

* Refs #3447. Modified discovery examples

* Refs #3447. Added ref integration tests.

* Discretize transport by parameter.

* Refs #3447. Updated with changes in develop

* Fixed bug at info cross_serialization test.

* Refs #3447. Updated discovery integration test.

* Modified xml of integration test to fix bug.

* Add reuse addr to tcp transport.

* Refs #3526. Refactor input serial transport.

* Refs #3526. Refactor output serial transport.

* Refs #3526. Added new serial tests.

* Refs #3550. Refactor Serial transport stack and modified transport folders.

* Refs #3526. Simplified Serial transport protocol.

* Refs #3526. Solved rebase conflicts.

* Refs #3550. Refactor TCP transport stack.

* Refs #3550. Refactor UDP transport stack.

* Refs #3550. Moved main Serial transport functions.

* Refs #3550. Added simple error management.

* Updated and fixed deployment example.

+++ src/c/core/serialization/xrce_protocol_internal.h
@@ -49,7 +49,6 @@ typedef struct BinarySequence_t
 
 } BinarySequence_t;
 
-
 typedef struct StringSequence_t
 {
     uint32_t size;
@@ -57,12 +56,12 @@ typedef struct StringSequence_t
 
 } StringSequence_t;
 
-
 typedef struct ClientKey
 {
     uint8_t data[4];
 
 } ClientKey;
+
 #define CLIENT_INVALID (ClientKey){{0x00, 0x00, 0x00, 0x00}}
 
 typedef uint8_t ObjectKind;
@@ -145,7 +144,7 @@ typedef struct TransportLocatorSmall
 typedef struct TransportLocatorMedium
 {
     uint8_t address[4];
-    uint8_t locator_port;
+    uint16_t locator_port;
 
 } TransportLocatorMedium;
 
@@ -153,7 +152,7 @@ typedef struct TransportLocatorMedium
 typedef struct TransportLocatorLarge
 {
     uint8_t address[16];
-    uint8_t locator_port;
+    uint32_t locator_port;
 
 } TransportLocatorLarge;
